# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "usuario@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "SenhaSegura123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "valido_1": {
                  "summary": "Login válido (1)",
                  "value": {
                    "email": "usuario@example.com",
                    "password": "SenhaSegura123"
                  }
                },
                "valido_2": {
                  "summary": "Login válido (2)",
                  "value": {
                    "email": "usuario@example.com",
                    "password": "SenhaSegura123"
                  }
                },
                "valido_3": {
                  "summary": "Login válido Admin",
                  "value": {
                    "email": "testeStudio@example.com",
                    "password": "123456"
                  }
                },
                "invalido_email_invalido": {
                  "summary": "Email em formato inválido",
                  "value": {
                    "email": "usuario",
                    "password": "minhasenha"
                  },
                  "description": "❌ Falha: email não é um endereço válido."
                },
                "invalido_senha_curta": {
                  "summary": "Senha com menos de 8 caracteres",
                  "value": {
                    "email": "teste@example.com",
                    "password": "123"
                  },
                  "description": "❌ Falha: password tem menos de 8 caracteres."
                },
                "invalido_email_vazio": {
                  "summary": "Email vazio",
                  "value": {
                    "email": "",
                    "password": "abcdefgh"
                  },
                  "description": "❌ Falha: email vazio não é válido."
                },
                "invalido_senha_vazia": {
                  "summary": "Senha vazia",
                  "value": {
                    "email": "user@site.com",
                    "password": ""
                  },
                  "description": "❌ Falha: password vazio não atende ao MinLength(8)."
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/perfil": {
      "get": {
        "operationId": "AuthController_getPerfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "API com Swagger",
    "description": "Documentação automática da API com Swagger",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "usuario@example.com",
            "description": "Endereço de e-mail único do usuário"
          },
          "name": {
            "type": "string",
            "example": "João Silva",
            "description": "Nome completo do usuário"
          },
          "password": {
            "type": "string",
            "example": "SenhaSegura123",
            "description": "Senha do usuário"
          },
          "age": {
            "type": "number",
            "example": 25,
            "description": "Idade do usuário (opcional)"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      }
    }
  }
}